Initial conversion of brltty to make it Unicode-based. (dm,st)
git-svn-id: svn://mielke.cc/main/brltty@3472 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ SpeechDrivers/BrailleLite/speech.c
@@ -64,7 +64,7 @@ spk_construct (SpeechSynthesizer *spk, char **parameters)
 
 
 static void
-spk_say (SpeechSynthesizer *spk, const unsigned char *buffer, int len)
+spk_say (SpeechSynthesizer *spk, const unsigned char *buffer, size_t len)
 {
   static unsigned char pre_speech[] = { PRE_SPEECH };
   static unsigned char post_speech[] = { POST_SPEECH };
