+++ SpeechDrivers/ViaVoice/speech.c
@@ -615,7 +615,7 @@ saySegment (ECIHand eci, const unsigned char *buffer, int from, int to) {
 }
 
 static void
-spk_say (SpeechSynthesizer *spk, const unsigned char *buffer, int length) {
+spk_say (SpeechSynthesizer *spk, const unsigned char *buffer, size_t length) {
    if (eci) {
       int onSpace = -1;
       int sayFrom = 0;
